+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Update Go to v1.19 (@timoreimann)
 * Support specifying region explicitly (@shatoboar)
 * Support custom annotation to specify HTTP3 entry ports for Load Balancers (@anitgandhi)
+* Support custom annotation to specify Project ID for Load Balancers (@StephenVarela)
 
 ## v0.1.39 (beta) - August 17, 2022
 * Updates kubernetes dependencies: